# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateTransactionRequest:
    properties:
      amount:
        type: string
      note:
        type: string
      t_date:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
    required:
    - amount
    - t_date
    type: object
...
  model.PurgeTransactionsRequest:
    properties:
      cutoff_date:
        type: string
    required:
    - cutoff_date
    type: object
...
  model.UpdateTransactionRequest:
    properties:
      deleted:
        type: boolean
      note:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A REST API for personal budget management with transaction tracking,
    recurring payments, and reporting.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Budget API
  version: "1.0"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Get all transactions for the authenticated user, optionally filtered
        by date range
      parameters:
      - description: Start date (YYYY-MM-DD format)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD format)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid date format
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction with optional tag associations
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /transactions/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transaction ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
    patch:
      consumes:
      - application/json
      description: Update an existing transaction's note, tags, or soft delete it
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated successfully
          schema:
            additionalProperties: true
            type: object
        "204":
          description: Transaction deleted successfully
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a transaction
      tags:
      - transactions
  /transactions/by-recurring/{recurring_id}:
    get:
      consumes:
      - application/json
      description: Get all transactions that were created from a specific recurring
        rule
      parameters:
      - description: Recurring rule ID
        in: path
        name: recurring_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid recurring ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get transactions by recurring ID
      tags:
      - transactions
  /transactions/by-tag/{tag_id}:
    get:
      consumes:
      - application/json
      description: Get all transactions associated with a specific tag
      parameters:
      - description: Tag ID
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid tag ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get transactions by tag
      tags:
      - transactions
  /transactions/purge:
    post:
      consumes:
      - application/json
      description: Permanently delete transactions that were soft deleted before a
        specified date
      parameters:
      - description: Purge request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PurgeTransactionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Purge completed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Purge soft deleted transactions
      tags:
      - transactions
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"
...